+++ skaffold-operator.yaml
@@ -24,6 +24,7 @@ build:
             - cmd/gardener-operator/app
             - cmd/gardener-operator/app/bootstrappers
             - cmd/utils
+            - cmd/utils/initrun
             - extensions/pkg/apis/config
             - extensions/pkg/controller
             - extensions/pkg/predicate
@@ -328,6 +329,7 @@ build:
             - cmd/gardener-resource-manager/app
             - cmd/gardener-resource-manager/app/bootstrappers
             - cmd/utils
+            - cmd/utils/initrun
             - pkg/api/extensions
             - pkg/api/indexer
             - pkg/apis/core
@@ -658,6 +660,7 @@ build:
             - cmd/gardener-controller-manager/app
             - cmd/gardener-controller-manager/app/bootstrappers
             - cmd/utils
+            - cmd/utils/initrun
             - pkg/api/extensions
             - pkg/api/indexer
             - pkg/apis/core
@@ -772,6 +775,7 @@ build:
             - cmd/gardener-scheduler
             - cmd/gardener-scheduler/app
             - cmd/utils
+            - cmd/utils/initrun
             - pkg/api/extensions
             - pkg/apis/core
             - pkg/apis/core/install
@@ -841,6 +845,7 @@ build:
             - cmd/gardener-admission-controller
             - cmd/gardener-admission-controller/app
             - cmd/utils
+            - cmd/utils/initrun
             - pkg/admissioncontroller/apis/config
             - pkg/admissioncontroller/apis/config/helper
             - pkg/admissioncontroller/apis/config/v1alpha1
